refactor(e2e): share plugin path containment checks
