+++ scripts/e2e/lib/openclaw-state-paths.mjs
@@ -1,4 +1,16 @@
+import fs from "node:fs";
 import path from "node:path";
+
+export function assertRealPathInside(parentPath, childPath, label) {
+  const parentRealPath = fs.realpathSync(parentPath);
+  const childRealPath = fs.realpathSync(childPath);
+  if (
+    childRealPath !== parentRealPath &&
+    !childRealPath.startsWith(`${parentRealPath}${path.sep}`)
+  ) {
+    throw new Error(`${label} resolved outside ${parentPath}: ${childRealPath}`);
+  }
+}
 
 export function resolveHomePath(value) {
   if (value === "~") {
